refactor: Simplify crawl and job commands with centralized app lifecycle management
- Extract common signal handling and graceful shutdown logic to `internal/common/app`
- Remove duplicate code for signal handling, context management, and application lifecycle
- Introduce `WaitForSignal()`, `SetupCollector()`, and `ConfigureCrawler()` utility functions
- Streamline command execution with consistent shutdown and error handling
- Improve code readability and reduce boilerplate in crawl and job commands
