+++ cmd/crawl/crawl.go
@@ -5,14 +5,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"os"
-	"os/signal"
-	"syscall"
-	"time"
 
 	"github.com/jonesrussell/gocrawl/internal/article"
 	"github.com/jonesrussell/gocrawl/internal/collector"
 	"github.com/jonesrussell/gocrawl/internal/common"
+	"github.com/jonesrussell/gocrawl/internal/common/app"
 	"github.com/jonesrussell/gocrawl/internal/config"
 	"github.com/jonesrussell/gocrawl/internal/content"
 	"github.com/jonesrussell/gocrawl/internal/crawler"
@@ -21,12 +18,6 @@ import (
 	"github.com/jonesrussell/gocrawl/internal/sources"
 	"github.com/spf13/cobra"
 	"go.uber.org/fx"
-)
-
-const (
-	// DefaultChannelBufferSize is the default size for buffered channels used for
-	// processing articles and content during crawling.
-	DefaultChannelBufferSize = 100
 )
 
 // sourceName holds the name of the source being crawled, populated from command line arguments.
@@ -38,8 +29,6 @@ func SetSourceName(name string) {
 }
 
 // Params holds the dependencies and parameters required for the crawl operation.
-// It uses fx.In to indicate that these fields should be injected by the fx dependency
-// injection framework.
 type Params struct {
 	fx.In
 
@@ -53,7 +42,6 @@ type Params struct {
 	CrawlerInstance crawler.Interface
 
 	// Processors is a slice of content processors, injected as a group
-	// The first processor handles articles, the second handles content
 	Processors []models.ContentProcessor `group:"processors"`
 
 	// Logger provides structured logging capabilities
@@ -69,9 +57,38 @@ type Params struct {
 	Context context.Context `name:"crawlContext"`
 }
 
-// Cmd represents the crawl command that initiates the crawling process for a
-// specified source. It reads the source configuration from sources.yml and starts
-// the crawling process with the configured parameters.
+// StartCrawl initializes and starts the crawling process.
+func StartCrawl(p Params) error {
+	return startCrawl(p)
+}
+
+// startCrawl is the internal implementation of StartCrawl
+func startCrawl(p Params) error {
+	if p.CrawlerInstance == nil {
+		return errors.New("crawler is not initialized")
+	}
+
+	// Get the source configuration first
+	source, err := p.Sources.FindByName(sourceName)
+	if err != nil {
+		return fmt.Errorf("error finding source: %w", err)
+	}
+
+	// Set up the collector
+	collectorResult, err := app.SetupCollector(p.Context, p.Logger, *source, p.Processors, p.Done, p.Config)
+	if err != nil {
+		return fmt.Errorf("error setting up collector: %w", err)
+	}
+
+	// Configure the crawler
+	if err := app.ConfigureCrawler(p.CrawlerInstance, *source, collectorResult); err != nil {
+		return fmt.Errorf("error configuring crawler: %w", err)
+	}
+
+	return nil
+}
+
+// Cmd represents the crawl command.
 var Cmd = &cobra.Command{
 	Use:   "crawl [source]",
 	Short: "Crawl a single source defined in sources.yml",
@@ -96,16 +113,12 @@ Example:
 		ctx, cancel := context.WithCancel(cmd.Context())
 		defer cancel()
 
-		// Set up signal handling for graceful shutdown
-		sigChan := make(chan os.Signal, 1)
-		signal.Notify(sigChan, syscall.SIGINT, syscall.SIGTERM)
-
-		// Create channels for error handling and completion
-		errChan := make(chan error, 1)
-		doneChan := make(chan struct{})
+		// Set up signal handling
+		done, cleanup := app.WaitForSignal(ctx, cancel)
+		defer cleanup()
 
 		// Initialize the Fx application with required modules and dependencies
-		app := fx.New(
+		fxApp := fx.New(
 			common.Module,
 			article.Module,
 			content.Module,
@@ -114,7 +127,7 @@ Example:
 			fx.Provide(
 				fx.Annotate(
 					func() chan struct{} {
-						return doneChan
+						return make(chan struct{})
 					},
 					fx.ResultTags(`name:"crawlDone"`),
 				),
@@ -141,207 +154,25 @@ Example:
 					fx.ResultTags(`name:"contentIndex"`, `name:"indexName"`),
 				),
 				func() chan *models.Article {
-					return make(chan *models.Article, DefaultChannelBufferSize)
+					return make(chan *models.Article, app.DefaultChannelBufferSize)
 				},
 			),
-			fx.Invoke(func(lc fx.Lifecycle, p Params) {
-				configureCrawlLifecycle(lc, p, errChan)
-			}),
 		)
 
-		// Start the application and handle any startup errors
-		if err := app.Start(ctx); err != nil {
+		// Start the application
+		if err := fxApp.Start(ctx); err != nil {
 			return fmt.Errorf("error starting application: %w", err)
 		}
 
-		// Wait for either:
-		// - A signal interrupt (SIGINT/SIGTERM)
-		// - Context cancellation
-		// - Crawl completion
-		// - Crawl error
-		var crawlErr error
-		select {
-		case sig := <-sigChan:
-			common.PrintInfof("\nReceived signal %v, initiating shutdown...", sig)
-			cancel() // Cancel our context
-		case <-ctx.Done():
-			common.PrintInfof("\nContext cancelled, initiating shutdown...")
-		case crawlErr = <-errChan:
-			// Error already printed in startCrawl
-			cancel() // Cancel our context on error
-		case <-doneChan:
-			// Success message already printed in startCrawl
-		}
+		// Wait for completion or interruption
+		<-done
 
-		// Create a context with timeout for graceful shutdown
-		stopCtx, stopCancel := context.WithTimeout(context.Background(), common.DefaultShutdownTimeout)
-		defer stopCancel()
-
-		// Stop the application and handle any shutdown errors
-		if err := app.Stop(stopCtx); err != nil && !errors.Is(err, context.Canceled) {
-			common.PrintErrorf("Error stopping application: %v", err)
-			return err
-		}
-
-		return crawlErr
+		// Perform graceful shutdown
+		return app.GracefulShutdown(fxApp)
 	},
 }
 
-// StartCrawl initializes and starts the crawling process. It sets up the crawler
-// with the specified source configuration and manages the crawl lifecycle.
-//
-// Parameters:
-//   - p: Params containing all required dependencies and configuration
-//
-// Returns:
-//   - error: Any error that occurred during setup or initialization
-func StartCrawl(p Params) error {
-	return startCrawl(p)
-}
-
-// startCrawl is the internal implementation of StartCrawl
-func startCrawl(p Params) error {
-	if p.CrawlerInstance == nil {
-		return errors.New("crawler is not initialized")
-	}
-
-	// Get the source configuration first
-	source, err := p.Sources.FindByName(sourceName)
-	if err != nil {
-		return fmt.Errorf("error finding source: %w", err)
-	}
-
-	// Convert the source config to the expected type
-	sourceConfig := common.ConvertSourceConfig(source)
-	if sourceConfig == nil {
-		return errors.New("source configuration is nil")
-	}
-
-	// Parse and validate rate limit from configuration
-	rateLimit, err := time.ParseDuration(source.RateLimit)
-	if err != nil {
-		return fmt.Errorf("invalid rate limit: %w", err)
-	}
-
-	// Extract domain from source URL
-	domain, err := common.ExtractDomain(source.URL)
-	if err != nil {
-		return fmt.Errorf("error extracting domain: %w", err)
-	}
-
-	// Create and configure the collector
-	collectorResult, err := collector.New(collector.Params{
-		BaseURL:          source.URL,
-		MaxDepth:         source.MaxDepth,
-		RateLimit:        rateLimit,
-		Debugger:         logger.NewCollyDebugger(p.Logger),
-		Logger:           p.Logger,
-		ArticleProcessor: p.Processors[0], // First processor handles articles
-		ContentProcessor: p.Processors[1], // Second processor handles content
-		Source:           sourceConfig,
-		Parallelism:      p.Config.GetCrawlerConfig().Parallelism,
-		RandomDelay:      p.Config.GetCrawlerConfig().RandomDelay,
-		Context:          p.Context,
-		Done:             p.Done,
-		AllowedDomains:   []string{domain},
-	})
-	if err != nil {
-		return fmt.Errorf("error creating collector: %w", err)
-	}
-
-	// Set the collector in the crawler instance
-	p.CrawlerInstance.SetCollector(collectorResult.Collector)
-
-	// Set crawler configuration
-	p.CrawlerInstance.SetMaxDepth(source.MaxDepth)
-	if rateLimitErr := p.CrawlerInstance.SetRateLimit(source.RateLimit); rateLimitErr != nil {
-		return fmt.Errorf("error setting rate limit: %w", rateLimitErr)
-	}
-
-	return nil
-}
-
-// initializeCrawl initializes the crawler with the given source
-func initializeCrawl(_ context.Context, p Params) (*sources.Config, error) {
-	source, findErr := p.Sources.FindByName(sourceName)
-	if findErr != nil {
-		return nil, fmt.Errorf("error finding source: %w", findErr)
-	}
-
-	if err := startCrawl(p); err != nil {
-		return nil, fmt.Errorf("failed to initialize crawler: %w", err)
-	}
-
-	return source, nil
-}
-
-// runCrawl executes the crawl operation in a goroutine
-func runCrawl(ctx context.Context, p Params, source *sources.Config, errChan chan error) {
-	go func() {
-		p.Logger.Info("Starting crawl", "source", sourceName)
-		if startErr := p.CrawlerInstance.Start(ctx, source.URL); startErr != nil {
-			if !errors.Is(startErr, context.Canceled) {
-				p.Logger.Error("Crawl failed", "error", startErr)
-				errChan <- startErr
-			}
-			close(p.Done)
-			return
-		}
-
-		p.CrawlerInstance.Wait()
-		p.Logger.Info("Crawl completed successfully", "source", sourceName)
-		close(p.Done)
-	}()
-}
-
-// handleShutdown manages the graceful shutdown of the crawler
-func handleShutdown(ctx context.Context, p Params) error {
-	p.Logger.Info("Stopping crawler...")
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	case <-p.Done:
-		p.Logger.Info("Crawl completed, stopping gracefully")
-		if stopErr := p.CrawlerInstance.Stop(ctx); stopErr != nil {
-			p.Logger.Error("Error stopping crawler", "error", stopErr)
-		}
-		return nil
-	default:
-		p.Logger.Info("Forcing crawler to stop")
-		if stopErr := p.CrawlerInstance.Stop(ctx); stopErr != nil {
-			p.Logger.Error("Error stopping crawler", "error", stopErr)
-		}
-		return nil
-	}
-}
-
-// configureCrawlLifecycle configures the lifecycle hooks for crawl management.
-// It sets up the start and stop hooks for the crawl operation.
-func configureCrawlLifecycle(lc fx.Lifecycle, p Params, errChan chan error) {
-	lc.Append(fx.Hook{
-		OnStart: func(ctx context.Context) error {
-			source, err := initializeCrawl(ctx, p)
-			if err != nil {
-				return err
-			}
-
-			runCrawl(ctx, p, source, errChan)
-
-			// Wait for either error or completion
-			select {
-			case channelErr := <-errChan:
-				return channelErr
-			case <-p.Done:
-				return nil
-			}
-		},
-		OnStop: func(ctx context.Context) error {
-			return handleShutdown(ctx, p)
-		},
-	})
-}
-
-// Command returns the crawl command for use in the root command
+// Command returns the crawl command.
 func Command() *cobra.Command {
 	return Cmd
 }
